+++ src/test/java/uk/ac/ebi/eva/pipeline/io/VepProcessTest.java
@@ -77,8 +77,8 @@ public class VepProcessTest {
     private String getVariantInVepInputFormat(EnsemblVariant ensemblVariant) {
         return String.join("\t",
                            ensemblVariant.getChr(),
-                           Integer.toString(ensemblVariant.getStart()),
-                           Integer.toString(ensemblVariant.getEnd()),
+                           Long.toString(ensemblVariant.getStart()),
+                           Long.toString(ensemblVariant.getEnd()),
                            ensemblVariant.getRefAlt(),
                            ensemblVariant.getStrand());
     }
