+++ src/test/java/uk/ac/ebi/eva/pipeline/model/EnsemblVariantTest.java
@@ -79,7 +79,7 @@ public class EnsemblVariantTest {
         List<Variant> result = factory.create(FILE_ID, STUDY_ID, line);
         assertEquals(1, result.size());
 
-        int start = result.get(0).getStart();
+        long start = result.get(0).getStart();
         String reference = result.get(0).getReference();
         String alternate = result.get(0).getAlternate();
 
@@ -100,7 +100,7 @@ public class EnsemblVariantTest {
         List<Variant> result = factory.create(FILE_ID, STUDY_ID, line);
         assertEquals(1, result.size());
 
-        int start = result.get(0).getStart();
+        long start = result.get(0).getStart();
         String reference = result.get(0).getReference();
         String alternate = result.get(0).getAlternate();
